Function for downloading CHELSA CMIP 5 future climatologies for the years 2041-2060 and 2061-2080

This function supports the download of CHELSA CMIP5 future climate scenarios comprising of monthly precipitation sums in mm, monthly temperature (average, minimum, maximum) in degrees Celsius, and annual chracteristics (19 bioclimatic variables).
To allow pre-processing, clipping and buffering, conversion to ASCII-grids and stacking options are included.
Optional an output of a .bib-file of the cited literature can be retrieved.
For user convenience, saving directories will be created automatically. Also options to "zip" and/or delete the RAW-files are included.

Chelsa.CMIP_5.download(
  save.location = "./",
  parameter = c("prec", "temp", "tmax", "tmin", "bio"),
  bio.var = c(1:19),
  month.var = c(1:12),
  emission.scenario.var = c("rcp26", "rcp45", "rcp60", "rcp85"),
  time.interval.var = c("2041-2060", "2061-2080"),
  model.var = c("ACCESS1-0", "bcc-csm1-1", "BNU-ESM", "CanESM2", "CCSM4", "CESM1-BGC",
    "CESM1-CAM5", "CMCC-CESM", "CMCC-CM", "CMCC-CMS", "CNRM-CM5", "CSIRO-Mk3-6-0",
    "CSIRO-Mk3L-1-2", "EC-EARTH", "FGOALS-g2", "FIO-ESM", "GFDL-CM3", "GFDL-ESM2G",
    "GFDL-ESM2M", "GISS-E2-H", "GISS-E2-H-CC", "GISS-E2-R", "GISS-E2-R-CC", "HadGEM2-AO",
    "HadGEM2-CC", "HadGEM2-ES", "inmcm4", "IPSL-CM5A-LR", "IPSL-CM5A-MR", "MIROC-ESM",
    "MIROC-ESM-CHEM", "MIROC5", "MPI-ESM-LR", "MPI-ESM-MR", "MRI-CGCM3", "MRI-ESM1",
    "NorESM1-M", "NorESM1-ME"),
  clipping = FALSE,
  clip.shapefile = NULL,
  clip.extent = c(-180, 180, -90, 90),
  buffer = 0,
  convert.files.to.asc = FALSE,
  stacking.data = FALSE,
  combine.raw.zip = FALSE,
  delete.raw.data = FALSE,
  save.bib.file = TRUE
)

Arguments

save.location

string. Input where the datasets should be saved.
Default: Working Directory.

parameter

string (vector). Input of parameters which should be downloaded.
Default: c("prec", "temp", "tmax", "tmin", "bio")

bio.var

integer (vector). Input which monthly data should be downloaded. Only applicable to BIOCLIM variables. For further information see: http://chelsa-climate.org/bioclim/.
Default: c(1:19)

month.var

integer (vector). Input which monthly data should be downloaded. Only applicable to Precipitation and Temperature (average, maximum, minimum).
Default: c(1:12)

emission.scenario.var

string (vector). Input which emission scenario dataset should be downloaded. Provided are the representative concentration pathways (RCP) 2.6, 4.5, 6.0, and 8.5.
Default: c("rcp26", "rcp45", "rcp60", "rcp85")

time.interval.var

string (vector). Input for which time interval data should be downloaded. CHELSA provides downscaled CMIP5 climatologies for 2050 and 2070. Multiple inputs possible.
Default: c("2041-2060", "2061-2080")

model.var

string (vector). Input which future model dataset should be downloaded. For more information see: http://chelsa-climate.org/future/.
For some of the datasets not all downloads are available. For the ones that are not supported the data will not be downloaded and a warning will be prompted. For an overview please try "warnings()" after execution.
Default: c("ACCESS1-0", "bcc-csm1-1", "BNU-ESM", "CanESM2", "CCSM4", "CESM1-BGC",
"CESM1-CAM5", "CMCC-CESM", "CMCC-CM", "CMCC-CMS", "CNRM-CM5", "CSIRO-Mk3-6-0",
"CSIRO-Mk3L-1-2", "EC-EARTH", "FGOALS-g2", "FIO-ESM", "GFDL-CM3", "GFDL-ESM2G",
"GFDL-ESM2M","GISS-E2-H", "GISS-E2-H-CC", "GISS-E2-R", "GISS-E2-R-CC", "HadGEM2-AO",
"HadGEM2-CC", "HadGEM2-ES", "inmcm4", "IPSL-CM5A-LR", "IPSL-CM5A-MR","MIROC-ESM",
"MIROC-ESM-CHEM","MIROC5", "MPI-ESM-LR", "MPI-ESM-MR", "MRI-CGCM3", "MRI-ESM1",
"NorESM1-M","NorESM1-ME")

clipping

logical. Input whether the downloaded data should be clipped.
If FALSE clip.shapefile, buffer, clip.extent will be ignored.
Default: FALSE

clip.shapefile

string. Input which shapefile should be used for clipping.
Default: NULL

clip.extent

numeric (vector). Input vector with four numeric values. This is following the input order c("xleft", "xright", "ybottom", "ytop").
Default: c(-180, 180, -90, 90)

buffer

numeric. Input of decimal degrees of buffer around the shapefile and/or extent.
Default: 0

convert.files.to.asc

logical. Input whether files should be converted into the ASCII format.
If TRUE: a new subdirectory is created and the rawdata is saved there. If clipping is TRUE: the clipped raster files are also saved as ASCII grids.
Default: FALSE

stacking.data

logical. Input whether the downloaded data should be stacked as a netCDF-rasterstack.
Default: FALSE

combine.raw.zip

logical. Should the downloaded raw-data be "zipped".
Default: FALSE

delete.raw.data

logical. Should the downloaded raw-data be deleted. If the combine.raw.zip is TRUE: raw-data is still available in the zipped file.
Default: FALSE

save.bib.file

logical. Whether a BibTex-citation file of the dataset should be provided in the Working directory.
Default: TRUE

Value

Downscaled CHELSA CMIP5 climatologies for 2050 and 2070.

Details

"The downscaled data has been produced using climatological aided interpolation based on the 1979-2013 reference climatologies from CHELSA." (CHELSA Climate 2020: http://chelsa-climate.org/future/)

Note

Please note that the downloaded data for temperature and the therefore also the first eleven bioclim-variables are processed to °C with one significant decimal without offset and factor. Processing and conversion to other file-formats on a global dataset may take some time.
For some of the datasets not all models and rcps are available. For the ones that are not supported the data will not be downloaded and a warning will be prompted. See parameter model.var for more information or check the website of CHELSA Climate (http://chelsa-climate.org/future/). Please note, that the downloaded data for temperature and the therefore also the first eleven bioclim-variables are processed to °C without offset and factor. Processing and conversion to other file-formats on a global dataset may take some time.

References

D. N. Karger, O. Conrad, J. B\"ohner , et al. "Climatologies at high resolution for the earth's land surface areas". In: _Scientific Data_ 4.1 (Sep. 2017). DOI: 10.1038/sdata.2017.122. <URL: https://doi.org/10.1038/sdata.2017.122>.

D. N. Karger, O. Conrad, J. B\"ohner , et al. _Data from: Climatologies at high resolution for the earth's land surface areas_. En. 2018. DOI: 10.5061/DRYAD.KD1D4. <URL: http://datadryad.org/stash/dataset/doi:10.5061/dryad.kd1d4>.

Author

Helge Jentsch

Examples

if (FALSE) {
# Bioclim
Chelsa.CMIP_5.download(parameter = "bio",
                        bio.var = c(1,19),
                        emission.scenario.var = "rcp26",
                        time.interval.var = "2041-2060",
                        model.var = "MPI-ESM-LR")
# Precipitation
Chelsa.CMIP_5.download(parameter = "prec",
                        month.var = c(1,12),
                        emission.scenario.var = "rcp26",
                        time.interval.var = "2041-2060",
                        model.var = "MPI-ESM-LR")
}