Function for downloading the WorldClim v1.4 CMIP5 future climate dataset

This function supports the download, pre-processing and management of the WorldClim v1.4 CMIP5 future climate dataset.
This comprises of monthly precipitation sums (mm), temperature (maximum, minimum; °C), and annual chracteristics (19 bioclimatic variables).
To allow pre-processing, clipping and buffering, conversion to ASCII-grids and stacking options are included.
Optional an output of a .bib-file of the cited literature can be retrieved.
For user convenience, saving directories will be created automatically. Also options to "zip" and/or delete the RAW-files are included.

WorldClim.CMIP_5.download(
  save.location = "./",
  parameter = c("prec", "tmax", "tmin", "bio"),
  bio.var = c(1:19),
  month.var = c(1:12),
  resolution = c("10min", "5min", "2.5min", "30s"),
  model.var = c("ACCESS1-0", "BCC-CSM1-1", "CCSM4", "CESM1-CAM5-1-FV2", "CNRM-CM5",
    "GFDL-CM3", "GFDL-ESM2G", "GISS-E2-R", "HadGEM2-AO", "HadGEM2-CC", "HadGEM2-ES",
    "INMCM4", "IPSL-CM5A-LR", "MIROC-ESM-CHEM", "MIROC-ESM", "MIROC5", "MPI-ESM-LR",
    "MRI-CGCM3", "NorESM1-M"),
  emission.scenario.var = c("rcp26", "rcp45", "rcp60", "rcp85"),
  time.interval.var = c("2050", "2070"),
  clipping = FALSE,
  clip.shapefile = NULL,
  clip.extent = c(-180, 180, -90, 90),
  buffer = 0,
  convert.files.to.asc = FALSE,
  stacking.data = FALSE,
  keep.raw.zip = FALSE,
  delete.raw.data = FALSE,
  save.bib.file = TRUE
)

Arguments

save.location

string. Input where the datasets should be saved.
Default: Working Directory.

parameter

string (vector). Input of parameters which should be downloaded.
Default: c("prec", "tmax", "tmin", "bio")

bio.var

integer (vector). Input which monthly data should be downloaded. Only applicable to BIOCLIM variables. For further information see: https://www.worldclim.org/data/bioclim.html.
Default: c(1:19)

month.var

integer (vector). Input which monthly data should be downloaded. Only applicable to precipitation and temperature (maximum, minimum).
Default: c(1:12)

resolution

string (vector). Ranging from a 10 arc-minute resolution over 5 and 2.5 arc-minute to 30 arc-second resolution.
Default: c("10m", "5m", "2.5m", "30s")

model.var

string (vector). Model used to calculate the dataset.
Default: c("ACCESS1-0", "BCC-CSM1-1", "CCSM4", "CESM1-CAM5-1-FV2", "CNRM-CM5",
"GFDL-CM3", "GFDL-ESM2G", "GISS-E2-R", "HadGEM2-AO", "HadGEM2-CC", "HadGEM2-ES",
"INMCM4", "IPSL-CM5A-LR", "MIROC-ESM-CHEM", "MIROC-ESM", "MIROC5",
"MPI-ESM-LR","MRI-CGCM3", "NorESM1-M")

emission.scenario.var

string (vector). Input which emission scenario dataset should be downloaded. Provided are the representative concentration pathways (RCP) 2.6, 4.5, 6.0, and 8.5.
Default: c("rcp26", "rcp45", "rcp60", "rcp85")

time.interval.var

string (vector). Time interval for which the dataset is calculated. The given intervals are 2040-2060 (represented by "2050"), and 2060 to 2080 (represented by "2070").
Default: c("2050", "2070")

clipping

logical. Input whether the downloaded data should be clipped.
If FALSE; clip.shapefile, buffer, clip.extent will be ignored.
Default: FALSE

clip.shapefile

string. Input which shapefile should be used for clipping.
Default: NULL

clip.extent

numeric (vector). Input vector with four numeric values. This is following the input order c("xleft", "xright", "ybottom", "ytop").
Default: c(-180, 180, -90, 90)

buffer

numeric. Input of decimal degrees of buffer around the shapefile and/or extent.
Default: 0

convert.files.to.asc

logical. Input whether files should be converted into the ASCII format.
If TRUE a new subdirectory is created and the rawdata is saved there.
If clipping is TRUE the clipped raster files are also saved as ASCII grids.
Default: FALSE

stacking.data

logical. Input whether the downloaded data should be stacked as a netCDF-rasterstack.
Default: FALSE

keep.raw.zip

logical. Should the downloaded raw-data be "zipped".
Default: FALSE

delete.raw.data

logical. Should the downloaded raw-data be deleted.
If the combine.raw.zip-option is TRUE, raw-data is still available in the zipped file.
Default: FALSE

save.bib.file

logical. Whether a BibTex-citation file of the dataset should be provided in the Working directory.
Default: TRUE

Value

WorldClim 1.4 CMIP5 Future climate datasets for the periods of 2041-2060 and/or 2061-2080.

Note

Please note that this dataset is regarded to as "outdated" by the WorldClim creators. The download of the current dataset "CMIP6" is also provided by this package with the WorldClim.CMIP_6.download function.

Please note also that the downloaded data for temperature and the therefore also the first eleven bioclim-variables are processed to °C with one significant decimal without offset and factor. Processing and conversion to other file-formats on a global dataset may take some time depending on the spatial resolution.

References

R. J. Hijmans, S. E. Cameron, J. L. Parra, et al. "Very high resolution interpolated climate surfaces for global land areas". In: _International Journal of Climatology_ 25.15 (2005), pp. 1965-1978. DOI: 10.1002/joc.1276. <URL: https://doi.org/10.1002/joc.1276>.

Author

Helge Jentsch

Examples

if (FALSE) {
# Bioclim
WorldClim.CMIP_5.download(parameter = "bio",
                           bio.var = c(1,12),
                           resolution = "10min",
                           model.var = "MPI-ESM-LR",
                           emission.scenario.var = "rcp26",
                           time.interval.var = "2050")
# Precipitation
WorldClim.CMIP_5.download(parameter = "prec",
                           month.var = c(1,12),
                           resolution = "10min",
                           model.var = "MPI-ESM-LR",
                           emission.scenario.var = "rcp26",
                           time.interval.var = "2050")
}