Function for downloading the WorldClim v2.1 CMIP6 future climate dataset

This function supports the download, pre-processing and management of the WorldClim v2.1 CMIP6 future climate dataset.
This comprises of monthly precipitation sums (mm), temperature (maximum, minimum; °C), and annual chracteristics (19 bioclimatic variables).
To allow pre-processing, clipping and buffering, conversion to ASCII-grids and stacking options are included.
Optional an output of a .bib-file of the cited literature can be retrieved.
For user convenience, saving directories will be created automatically. Also options to "zip" and/or delete the RAW-files are included.

WorldClim.CMIP_6.download(
  save.location = "./",
  parameter = c("prec", "tmax", "tmin", "bio"),
  bio.var = c(1:19),
  month.var = c(1:12),
  resolution = c("10min", "5min", "2.5min", "30s"),
  model.var = c("BCC-CSM2-MR", "CNRM-CM6-1", "CNRM-ESM2-1", "CanESM5", "GFDL-ESM4",
    "IPSL-CM6A-LR", "MIROC-ES2L", "MIROC6", "MRI-ESM2-0"),
  emission.scenario.var = c("ssp126", "ssp245", "ssp370", "ssp585"),
  time.interval.var = c("2021-2040", "2041-2060", "2061-2080", "2081-2100"),
  clipping = FALSE,
  clip.shapefile = NULL,
  clip.extent = c(-180, 180, -90, 90),
  buffer = 0,
  convert.files.to.asc = FALSE,
  stacking.data = FALSE,
  keep.raw.zip = FALSE,
  delete.raw.data = FALSE,
  save.bib.file = TRUE
)

Arguments

save.location

string. Input where the datasets should be saved.
Default: Working Directory.

parameter

string (vector). Input of parameters which should be downloaded.
Default: c("prec", "tmax", "tmin", "bio")

bio.var

integer (vector). Input which monthly data should be downloaded. Only applicable to BIOCLIM variables. For further information see: https://www.worldclim.org/data/bioclim.html.
Default: c(1:19)

month.var

integer (vector). Input which monthly data should be downloaded. Only applicable to precipitation and temperature (maximum, minimum).
Default: c(1:12)

resolution

string (vector). Ranging from a 10 arc-minute resolution over 5 and 2.5 arc-minute to 30 arc-second resolution.
Default: c("10m", "5m", "2.5m", "30s")

model.var

string (vector). Model used to calculate the dataset.
Default: c("BCC-CSM2-MR", "CNRM-CM6-1", "CNRM-ESM2-1", "CanESM5", "GFDL-ESM4", "IPSL-CM6A-LR", "MIROC-ES2L", "MIROC6", "MRI-ESM2-0")

emission.scenario.var

string (vector). Input which emission scenario dataset should be downloaded. Provided are the shared socio-economic pathways (SSPs) 126, 245, 370, and 585.
Further information on SSPs: https://www.carbonbrief.org/cmip6-the-next-generation-of-climate-models-explained and https://www.carbonbrief.org/explainer-how-shared-socioeconomic-pathways-explore-future-climate-change.
Default: c("ssp126", "ssp245", "ssp370", "ssp585")

time.interval.var

string (vector). Averaged monthly values over a 20 year period. Periodes availabe are: 2021-2040, 2041-2060, 2061-2080, 2081-2100.
Default: c("2021-2040", "2041-2060", "2061-2080", "2081-2100")

clipping

logical. Input whether the downloaded data should be clipped.
If FALSE; clip.shapefile, buffer, clip.extent will be ignored.
Default: FALSE

clip.shapefile

string. Input which shapefile should be used for clipping.
Default: NULL

clip.extent

numeric (vector). Input vector with four numeric values. This is following the input order c("xleft", "xright", "ybottom", "ytop").
Default: c(-180, 180, -90, 90)

buffer

numeric. Input of decimal degrees of buffer around the shapefile and/or extent.
Default: 0

convert.files.to.asc

logical. Input whether files should be converted into the ASCII format.
If TRUE a new subdirectory is created and the rawdata is saved there. If "clipping" is TRUE the clipped raster files are also saved as ASCII grids.
Default: FALSE

stacking.data

logical. Input whether the downloaded data should be stacked as a netCDF-rasterstack.
Default: FALSE

keep.raw.zip

logical. Should the downloaded raw-data be "zipped".
Default: FALSE

delete.raw.data

logical. Should the downloaded raw-data be deleted. If the "combine.raw.zip"-option is TRUE, raw-data is still available in the zipped file.
Default: FALSE

save.bib.file

logical. Whether a BibTex-citation file of the dataset should be provided in the Working directory.
Default: TRUE

Value

WorldClim 2.1 CMIP6 Future climate datasets for the periods of 2021-2040, 2041-2060, 2061-2080 and/or 2081-2100.

Note

Please note that the downloaded data for temperature and the therefore also the first eleven bioclim-variables are processed to °C with one significant decimal without offset and factor. Processing and conversion to other file-formats on a global dataset may take some time depending on the spatial resolution.
**The 30 arc-second resolution is not yet available. It was scheduled to be released by March 2020 but until March 2021 no data was released publicly.**

References

S. E. Fick and R. J. Hijmans. "WorldClim 2: new 1-km spatial resolution climate surfaces for global land areas". In: _International Journal of Climatology_ 37.12 (Okt. 2017), pp. 4302-4315. DOI: 10.1002/joc.5086. <URL:https://doi.org/10.1002/joc.5086>.

Author

Helge Jentsch

Examples

if (FALSE) {
# Bioclim
WorldClim.CMIP_6.download(parameter = "bio",
                           bio.var = c(1,12),
                           resolution = "10min",
                           model.var = "MIROC6",
                           emission.scenario.var = "ssp126",
                           time.interval.var = "2021-2040")
# Precipitation
WorldClim.CMIP_6.download(parameter = "prec",
                           month.var = c(1,12),
                           resolution = "10min",
                           model.var = "MIROC6",
                           emission.scenario.var = "ssp126",
                           time.interval.var = "2021-2040")
}