Function for downloading the WorldClim historical climate dataset

This function supports the download, pre-processing and management of the WorldClim historical climate dataset.
This comprises of monthly precipitation sums (mm), temperature (average, maximum, minimum; °C), monthly solar radiation sums (kJ m-2 day-1), wind speed (m s-1), water vapor pressure (kPa), and annual chracteristics (19 bioclimatic variables). Also an elevation raster is provided.
To allow pre-processing, clipping and buffering, conversion to ASCII-grids and stacking options are included.
Optional an output of a .bib-file of the cited literature can be retrieved.
For user convenience, saving directories will be created automatically. Also options to "zip" and/or delete the RAW-files are included.

WorldClim.HistClim.download(
  save.location = "./",
  parameter = c("prec", "temp", "tmax", "tmin", "srad", "wind", "vapr", "bio", "elev"),
  bio.var = c(1:19),
  month.var = c(1:12),
  resolution = c("10min", "5min", "2.5min", "30s"),
  version.var = c("1.4", "2.1"),
  clipping = FALSE,
  clip.shapefile = NULL,
  clip.extent = c(-180, 180, -90, 90),
  buffer = 0,
  convert.files.to.asc = FALSE,
  stacking.data = FALSE,
  keep.raw.zip = FALSE,
  delete.raw.data = FALSE,
  save.bib.file = TRUE
)

Arguments

save.location

string. Input where the datasets should be saved.
Default: Working Directory.

parameter

string (vector). Input of parameters which should be downloaded.
Default: c("prec", "temp", "tmax","tmin", "srad", "wind", "vapr", "bio", "elev").

bio.var

integer (vector). Input which monthly data should be downloaded. Only applicable to BIOCLIM variables. For further information see: https://www.worldclim.org/data/bioclim.html.
Default: c(1:19)

month.var

integer (vector). Input which monthly data should be downloaded.
Default: c(1:12)

resolution

string (vector). Ranging from a 10 arc-minute resolution over 5 and 2.5 arc-minute to 30 arc-second resolution.
Default: c("10m", "5m", "2.5m", "30s")

version.var

string (vector). Input which version of the dataset should be downloaded. Multiple selection is possible.
Default: c("1.4", "2.1")

clipping

logical. Input whether the downloaded data should be clipped. See clipping.tif for more information.
If FALSE: clip.shapefile, buffer, clip.extent will be ignored.
Default: FALSE

clip.shapefile

string. Input which shapefile should be used for clipping.
Default: NULL

clip.extent

numeric (vector). Input vector with four numeric values. This is following the input order c("xleft", "xright", "ybottom", "ytop").
Default: c(-180, 180, -90, 90)

buffer

numeric. Input of decimal degrees of buffer around the shapefile and/or extent.
Default: 0

convert.files.to.asc

logical. Input whether files should be converted into the ASCII format. See convert.to.asc for more information.
If TRUE: a new subdirectory is created and the rawdata is saved there.
If the parameter clipping is also TRUE: the clipped raster files are also saved as ASCII grids.
Default: FALSE

stacking.data

logical. Input whether the downloaded data should be stacked as a netCDF-rasterstack. See stacking.downloaded.data for more information.
Default: FALSE

keep.raw.zip

logical. Should the downloaded raw-data be provided as "zip"-file. See combine.raw.in.zip for more information.
Default: FALSE

delete.raw.data

logical. Should the downloaded raw-data be deleted.
If the "combine.raw.zip"-option is TRUE: raw-data is still available in the zipped file.
Default: FALSE

save.bib.file

logical. Whether a BibTex-citation file of the dataset should be provided in the Working directory. See save.citation for more information.
Default: TRUE

Value

WorldClim climate datasets for the period of 1960-1990 (for v1.4) and/or 1970-2000 (for v2.1).

Note

Please note that solar radiation, wind speed, water vapor pressure, bioclimatic parameters, and elevation raster are only provided by the WorldClim Version 2.1 (current version).

Please note also that the downloaded data for temperature and the therefore also the first eleven bioclim-variables are processed to °C with one significant decimal without offset and factor. Processing and conversion to other file-formats on a global dataset may take some time depending on the spatial resolution.

References

R. J. Hijmans, S. E. Cameron, J. L. Parra, et al. "Very high resolution interpolated climate surfaces for global land areas". In: _International Journal of Climatology_ 25.15 (2005), pp. 1965-1978. DOI: 10.1002/joc.1276. <URL: https://doi.org/10.1002/joc.1276>.

S. E. Fick and R. J. Hijmans. "WorldClim 2: new 1-km spatial resolution climate surfaces for global land areas". In: _International Journal of Climatology_ 37.12 (Okt. 2017), pp. 4302-4315. DOI: 10.1002/joc.5086. <URL:https://doi.org/10.1002/joc.5086>.

Author

Helge Jentsch

Examples

if (FALSE) {
# Bioclim
WorldClim.HistClim.download(parameter = "bio",
                            bio.var = c(1,12),
                            resolution = "10min",
                            version.var = c("1.4", "2.1"))
# Precipitation
WorldClim.HistClim.download(parameter = "prec",
                             month.var = c(1,12),
                             resolution = "10min",
                             version.var = c("1.4", "2.1")
                             )
}