Combines all .tif-raster files into a .zip-file

Combines and saves all .tif-raster files to a .zip-file, whereas name and saving location can be specified.

combine.raw.in.zip(
  save.location = "./",
  zip.name = "RAWDATA",
  unique.name = TRUE,
  time.stamp.var = stringr::str_replace_all(stringr::str_replace_all(paste0(Sys.time()),
    pattern = ":", replacement = "-"), pattern = " ", replacement = "_")
)

Arguments

save.location

string (directory path). The directory where .tif-raster files are saved and the created/updated .zip file will be saved.
Default: "./" (Working directory)

zip.name

string. Input how the .zip-file should be named.
Default: "RAWDATA"

unique.name

logical. Should the .zip-file be named uniquely? If TRUE the current system time is added as a timestamp to create unique directories for multiple run outputs.
Default: TRUE

time.stamp.var

string. Input of current system time or, if called within another function the initial time of the execution.
Default: stringr::str_replace_all(stringr::str_replace_all(paste0(
Sys.time()),pattern = ":",replacement = "-"))

Author

Helge Jentsch