Converting .tif-raster files in one specified directory into ASCII-Grids

This function converts and saves all .tif-raster files to ASCII-file format to a specified location and creates a new folder named "ASCII_files" with a unique timestamp of the current system time.

convert.to.asc(
  save.location = "./",
  time.stamp.var = str_replace_all(str_replace_all(paste0(Sys.time()), pattern = ":",
    replacement = "-"), pattern = " ", replacement = "_")
)

Arguments

save.location

string (directory path). The directory where .tif-raster files are saved.
Default: "./" (Working directory)

time.stamp.var

string. Timestamp to create unique directories for multiple run outputs.
Default: stringr::str_replace_all(stringr::str_replace_all(paste0(
Sys.time()),pattern = ":",replacement = "-"))

Value

This function returns ASCII-format raster files to a new directory. This directory is dynamically created.

Author

Helge Jentsch

Examples

convert.to.asc(save.location = system.file("pictures/", package = "rgdal"))